# backend/src/main/java/tw/waterballsa/api/dto/response/CouponValidationResponse.java
package tw.waterballsa.api.dto.response;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.math.BigDecimal;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class CouponValidationResponse {
    private Boolean valid;
    private String code;
    private String discountType;
    private BigDecimal discountValue;
    private String errorCode;
    private String errorMessage;
}